+++ price-feeder/oracle/provider/huobi.go
@@ -267,8 +267,8 @@ func (p *HuobiProvider) setTickerPair(ticker HuobiTicker) {
 func (p *HuobiProvider) setCandlePair(candle HuobiCandle) {
 	p.mtx.Lock()
 	defer p.mtx.Unlock()
-	// huobi time period comes in seconds
-	candle.Tick.TimeStamp = candle.Tick.TimeStamp * 1000
+	// convert huobi timestamp seconds -> milliseconds
+	candle.Tick.TimeStamp = candle.Tick.TimeStamp * int64(time.Second/time.Millisecond)
 	staleTime := PastUnixTime(providerCandlePeriod)
 	candleList := []HuobiCandle{}
 	candleList = append(candleList, candle)